Unloading packages instead of detaching
From the R help on ns-load {base}:
"unloadNamespace can be used to force a name space to be unloaded. An error is
signaled if the name space is imported by other loaded name spaces. If defined,
a hook function .onUnload is run before removing the name space from the
internal registry. unloadNamespace will first detach a package of the same name
if one is on the path, thereby running a .Last.lib function in the package if
one is exported."